Add SelectCommand
